feat(03-01): add Room entities, DAOs, mappers, and DB migration v2->v3
- Fix Purchase.quantity from Double to Int (STATE.md blocker)
- Add CartItem and WishlistItem domain models
- Add CartItemEntity with composite PK (userId, productId) for cart_items table
- Add WishlistItemEntity with composite PK (userId, productId) for wishlist_items table
- Add CartItemDao with observeCartItems, observeUniqueProductCount, upsert, updateQuantity, deleteItem, clearCart, updateProductStatus
- Add WishlistItemDao with observeWishlistItems, isWishlisted, upsert, deleteItem, deleteAllForUser
- Add CartItemMapper and WishlistItemMapper (entity <-> domain)
- Update WenuCommerceDatabase to version 3 with MIGRATION_2_3 creating both new tables
- Export Room schema v3 JSON

+++ domain/src/main/java/com/wenubey/domain/model/Purchase.kt
@@ -7,7 +7,7 @@ data class Purchase(
     val purchaseId: String = "",
     val productId: String = "",
     val variantId: String = "",
-    val quantity: Double = 0.0,
+    val quantity: Int = 0,
     val price: Double = 0.0,
     val purchaseDate: String = "",
     val stripePaymentIntentId: String = "",
